+++ apps/console/src/extensions/i18n/resources/fr-FR/extensions.ts
@@ -1150,6 +1150,25 @@ export const extensions: Extensions = {
                                 }
                             }
                         }
+                    },
+                    rolesV2: {
+                        heading: "Les rôles",
+                        subHeading: "Gérer les rôles attribués dans l'application.",
+                        roleAudience: "Rôle public",
+                        organization: "Organisation",
+                        application: "application",
+                        assignedRoles: "Rôles attribués",
+                        removedRoles: "Rôles supprimés",
+                        searchPlaceholder: "Recherche par nom de rôle",
+                        switchRoleAudience: {
+                            confirmationModal: {
+                                assertionHint: "Veuillez confirmer votre action.",
+                                content: "Si vous modifiez le public du rôle, les rôles associés au public actuel " +
+                                    "être supprimé.Veuillez procéder avec prudence.",
+                                header: "Es-tu sûr?",
+                                message: "Cette action est irréversible et supprimera en permanence les rôles existants."
+                            }
+                        }
                     }
                 }
             },
@@ -1299,7 +1318,7 @@ export const extensions: Extensions = {
                                     "d'entre eux. S'il n'est pas défini, les valeurs par défaut de " +
                                     "{{ productName }} seront utilisées.",
                                 label: "Politique relative aux cookies",
-                                placeholder: "https://asgardeo.io/cookie-policy"
+                                placeholder: "https://myapp.com/cookie-policy"
                             },
                             privacyPolicyURL: {
                                 hint:
@@ -1308,7 +1327,7 @@ export const extensions: Extensions = {
                                     "organisation. S'il n'est pas défini, les valeurs par défaut de " +
                                     "{{ productName }} seront utilisées.",
                                 label: "Politique de confidentialité",
-                                placeholder: "https://asgardeo.io/privacy-policy"
+                                placeholder: "https://myapp.com/privacy-policy"
                             },
                             termsOfUseURL: {
                                 hint:
@@ -1317,7 +1336,7 @@ export const extensions: Extensions = {
                                     "vos clients et visiteurs. S'il n'est pas défini, les valeurs par défaut " +
                                     "de {{ productName }} seront utilisées.",
                                 label: "Conditions d'utilisation",
-                                placeholder: "https://asgardeo.io/terms-of-service"
+                                placeholder: "https://myapp.com/terms-of-service"
                             }
                         },
                         heading: "Liens"