+++ apps/console/src/features/core/configs/routes.tsx
@@ -150,35 +150,6 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
                         children: [
                             {
                                 component: lazy(() =>
-                                    import("../../../extensions/components/" + "api-resources/pages/api-resource-edit")
-                                ),
-                                exact: true,
-                                id: "apiResources-edit",
-                                name: "extensions:develop.sidePanel.apiResources",
-                                path: APIResourcesConstants.getPaths().get("API_RESOURCE_EDIT"),
-                                protected: true,
-                                showOnSidePanel: false
-                            }
-                        ],
-                        component: lazy(() =>
-                            import("../../../extensions/components/" + "api-resources/pages/api-resources")
-                        ),
-                        exact: true,
-                        icon: {
-                            icon: import("../../../extensions/assets/images/icons/api-resources-icon.svg")
-                        },
-                        id: "apiResources",
-                        name: "extensions:develop.sidePanel.apiResources",
-                        order: 2,
-                        path: APIResourcesConstants.getPaths().get("API_RESOURCES"),
-                        protected: true,
-                        showOnSidePanel: true
-                    },
-                    {
-                        category: "console:develop.features.sidePanel.categories.application",
-                        children: [
-                            {
-                                component: lazy(() =>
                                     import("../../connections/pages/connection-templates")
                                 ),
                                 exact: true,
@@ -264,36 +235,6 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
                         name: "console:develop.features.sidePanel.categories.identityVerificationProviders",
                         order: 4,
                         path: AppConstants.getPaths().get("IDVP"),
-                        protected: true,
-                        showOnSidePanel: true
-                    },
-                    {
-                        category: "extensions:manage.sidePanel.categories.userManagement",
-                        children: [
-                            {
-                                component: lazy(() => import("../../application-roles/pages/application-role-edit")),
-                                exact: true,
-                                icon: {
-                                    icon: getSidePanelIcons().childIcon
-                                },
-                                id: "applicationRolesEdit",
-                                name: "Edit Role",
-                                path: AppConstants.getPaths().get("APPLICATION_ROLES_EDIT"),
-                                protected: true,
-                                showOnSidePanel: false
-                            }
-                        ],
-                        component: lazy(() => import("../../application-roles/pages/application-roles")),
-                        exact: true,
-                        featureStatus: "NEW",
-                        featureStatusLabel: "common:new",
-                        icon: {
-                            icon: getSidePanelIcons().roles
-                        },
-                        id: "applicationRoles",
-                        name: "Roles",
-                        order: 8,
-                        path: AppConstants.getPaths().get("APPLICATION_ROLES"),
                         protected: true,
                         showOnSidePanel: true
                     },
@@ -521,7 +462,7 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
                         order: 17,
                         path: AppConstants.getPaths().get("MY_ACCOUNT"),
                         protected: true,
-                        showOnSidePanel: true
+                        showOnSidePanel: false
                     },
                     {
                         category: "extensions:manage.sidePanel.categories.AccountManagement",
@@ -812,7 +753,7 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
                         icon: {
                             icon: getSidePanelIcons().remoteLogging
                         },
-                        id: "remoteLogIngest",
+                        id: "remoteLogging",
                         name: "Remote Logging",
                         order: 22,
                         path: AppConstants.getPaths().get("REMOTE_LOGGING"),
@@ -955,6 +896,35 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
     if (useExtendedRoutes) {
         routes.push(
             {
+                category: "console:develop.features.sidePanel.categories.application",
+                children: [
+                    {
+                        component: lazy(() =>
+                            import("../../../extensions/components/api-resources/pages/api-resource-edit")
+                        ),
+                        exact: true,
+                        id: "apiResources-edit",
+                        name: "extensions:develop.sidePanel.apiResources",
+                        path: APIResourcesConstants.getPaths().get("API_RESOURCE_EDIT"),
+                        protected: true,
+                        showOnSidePanel: false
+                    }
+                ],
+                component: lazy(() =>
+                    import("../../../extensions/components/api-resources/pages/api-resources")
+                ),
+                exact: true,
+                icon: {
+                    icon: import("../../../extensions/assets/images/icons/api-resources-icon.svg")
+                },
+                id: "apiResources",
+                name: "extensions:develop.sidePanel.apiResources",
+                order: 2,
+                path: APIResourcesConstants.getPaths().get("API_RESOURCES"),
+                protected: true,
+                showOnSidePanel: true
+            },
+            {
                 category: "extensions:manage.sidePanel.categories.userManagement",
                 children: [
                     {
@@ -1055,6 +1025,36 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
                 name: "Groups",
                 order: 6,
                 path: AppConstants.getPaths().get("GROUPS"),
+                protected: true,
+                showOnSidePanel: true
+            },
+            {
+                category: "extensions:manage.sidePanel.categories.userManagement",
+                children: [
+                    {
+                        component: lazy(() => import("../../application-roles/pages/application-role-edit")),
+                        exact: true,
+                        icon: {
+                            icon: getSidePanelIcons().childIcon
+                        },
+                        id: "applicationRolesEdit",
+                        name: "Edit Role",
+                        path: AppConstants.getPaths().get("APPLICATION_ROLES_EDIT"),
+                        protected: true,
+                        showOnSidePanel: false
+                    }
+                ],
+                component: lazy(() => import("../../application-roles/pages/application-roles")),
+                exact: true,
+                featureStatus: "NEW",
+                featureStatusLabel: "common:new",
+                icon: {
+                    icon: getSidePanelIcons().roles
+                },
+                id: "applicationRoles",
+                name: "Roles",
+                order: 8,
+                path: AppConstants.getPaths().get("APPLICATION_ROLES"),
                 protected: true,
                 showOnSidePanel: true
             },
@@ -1194,6 +1194,35 @@ export const getAppViewRoutes = (useExtendedRoutes: boolean = false): RouteInter
         );
     } else {
         routes.push(
+            {
+                category: "console:develop.features.sidePanel.categories.application",
+                children: [
+                    {
+                        component: lazy(() =>
+                            import("../../api-resources/pages/api-resource-edit")
+                        ),
+                        exact: true,
+                        id: "apiResources-edit",
+                        name: "extensions:develop.sidePanel.apiResources",
+                        path: APIResourcesConstants.getPaths().get("API_RESOURCE_EDIT"),
+                        protected: true,
+                        showOnSidePanel: false
+                    }
+                ],
+                component: lazy(() =>
+                    import("../../api-resources/pages/api-resources")
+                ),
+                exact: true,
+                icon: {
+                    icon: import("../../../extensions/assets/images/icons/api-resources-icon.svg")
+                },
+                id: "apiResources",
+                name: "extensions:develop.sidePanel.apiResources",
+                order: 2,
+                path: APIResourcesConstants.getPaths().get("API_RESOURCES"),
+                protected: true,
+                showOnSidePanel: true
+            },
             {
                 category: "extensions:manage.sidePanel.categories.userManagement",
                 children: [