+++ apps/console/src/features/core/store/reducers/config.ts
@@ -125,8 +125,10 @@ export const commonConfigReducerInitialState: CommonConfigReducerStateInterface<
             requestPathAuthenticators: "",
             resourceTypes: "",
             roles: "",
+            rolesV2: "",
             rolesWithoutOrgPath: "",
             rootOrganization: "",
+            rootUsersOrganization: "",
             saml2Meta: "",
             schemas: "",
             selfSignUp: "",
@@ -139,6 +141,7 @@ export const commonConfigReducerInitialState: CommonConfigReducerStateInterface<
             userSessions: "",
             userStores: "",
             users: "",
+            usersOrganization: "",
             usersSuperOrganization: "",
             validationServiceMgt: "",
             wellKnown: ""