+++ apps/console/src/features/users/models/endpoints.ts
@@ -16,6 +16,8 @@
  * under the License.
  */
 
+import { UserDetailsInterface } from "./user";
+
 /**
  * Interface for the Users feature resource endpoints.
  */
@@ -46,7 +48,7 @@ export interface SCIMBulkOperation {
     method: string;
     bulkId: string;
     path: string;
-    data: Record<string, unknown>;
+    data: Record<string, unknown> | UserDetailsInterface;
 }
 
 /**