+++ apps/console/CHANGELOG.md
@@ -1,5 +1,87 @@
 # @wso2is/console
 
+## 2.1.10
+
+### Patch Changes
+
+- [#4260](https://github.com/wso2/identity-apps/pull/4260) [`ab4160125c`](https://github.com/wso2/identity-apps/commit/ab4160125cb6008de324e729d7af266cb2ed4bb9) Thanks [@dewniMW](https://github.com/dewniMW)! - Change API path that is used to fetch authorized orgs of the user
+
+## 2.1.9
+
+### Patch Changes
+
+- [#4229](https://github.com/wso2/identity-apps/pull/4229) [`eeb2da4a7c`](https://github.com/wso2/identity-apps/commit/eeb2da4a7c57d8870e6d1aef37a4337a151e1e6c) Thanks [@dasuni-30](https://github.com/dasuni-30)! - Implement the multiple user invites feature
+
+* [#4289](https://github.com/wso2/identity-apps/pull/4289) [`7997349f7a`](https://github.com/wso2/identity-apps/commit/7997349f7aeebf67e863a27f148e9c919a9c08f9) Thanks [@ThaminduDilshan](https://github.com/ThaminduDilshan)! - Enable back missing connection edit UI fields in JIT provisioning and Advanced config sections
+
+- [#4290](https://github.com/wso2/identity-apps/pull/4290) [`d32dfab48d`](https://github.com/wso2/identity-apps/commit/d32dfab48d2deb5ac94e5c056f018bb891b3b74b) Thanks [@Achintha444](https://github.com/Achintha444)! - Update the basic section of the role
+
+* [#4285](https://github.com/wso2/identity-apps/pull/4285) [`b99c4e0278`](https://github.com/wso2/identity-apps/commit/b99c4e0278624c0d9d9b373b236799e39b0f6829) Thanks [@SujanSanjula96](https://github.com/SujanSanjula96)! - Hide API Resource edit in side panel
+
+* Updated dependencies [[`eeb2da4a7c`](https://github.com/wso2/identity-apps/commit/eeb2da4a7c57d8870e6d1aef37a4337a151e1e6c), [`d32dfab48d`](https://github.com/wso2/identity-apps/commit/d32dfab48d2deb5ac94e5c056f018bb891b3b74b)]:
+  - @wso2is/i18n@2.0.12
+
+## 2.1.8
+
+### Patch Changes
+
+- [#4277](https://github.com/wso2/identity-apps/pull/4277) [`803278ca83`](https://github.com/wso2/identity-apps/commit/803278ca8368e8fa6a810473b22cb8e84559c7ed) Thanks [@DonOmalVindula](https://github.com/DonOmalVindula)! - Add application roles v2 UI
+
+* [#4278](https://github.com/wso2/identity-apps/pull/4278) [`8be0337845`](https://github.com/wso2/identity-apps/commit/8be03378456c5b8799fc3052e4b7a315b65f7013) Thanks [@Achintha444](https://github.com/Achintha444)! - Add permissions selection section to the create role interface
+
+- [#4271](https://github.com/wso2/identity-apps/pull/4271) [`5322255557`](https://github.com/wso2/identity-apps/commit/532225555766a62e4fcab31166e9bc9392a1584c) Thanks [@SujanSanjula96](https://github.com/SujanSanjula96)! - Modify API Resource UIs to support new API Resource Mgt Service
+
+* [#4280](https://github.com/wso2/identity-apps/pull/4280) [`7ee4e2e360`](https://github.com/wso2/identity-apps/commit/7ee4e2e360f463be9b9ca72814e4ee90077995da) Thanks [@Achintha444](https://github.com/Achintha444)! - Update roles list interface
+
+* Updated dependencies [[`8be0337845`](https://github.com/wso2/identity-apps/commit/8be03378456c5b8799fc3052e4b7a315b65f7013), [`7ee4e2e360`](https://github.com/wso2/identity-apps/commit/7ee4e2e360f463be9b9ca72814e4ee90077995da)]:
+  - @wso2is/i18n@2.0.11
+  - @wso2is/theme@2.0.18
+
+## 2.1.7
+
+### Patch Changes
+
+- [#4272](https://github.com/wso2/identity-apps/pull/4272) [`9e45286c38`](https://github.com/wso2/identity-apps/commit/9e45286c38484bffa96fc8a8d15c9b9f6249b37c) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - Update the route id for remote logging feature
+
+* [#4275](https://github.com/wso2/identity-apps/pull/4275) [`c34909a7c1`](https://github.com/wso2/identity-apps/commit/c34909a7c10c3f9cc9bde49dba118ea384f8ba9e) Thanks [@DilshanSenarath](https://github.com/DilshanSenarath)! - Fix the logo theme issue
+
+- [#4274](https://github.com/wso2/identity-apps/pull/4274) [`aa6e1e6012`](https://github.com/wso2/identity-apps/commit/aa6e1e6012fea5ec6c04d54f4c6e6b17e268239e) Thanks [@JayaShakthi97](https://github.com/JayaShakthi97)! - Add class for the `Reject All` button in the cookie banner
+
+- Updated dependencies [[`c34909a7c1`](https://github.com/wso2/identity-apps/commit/c34909a7c10c3f9cc9bde49dba118ea384f8ba9e), [`aa6e1e6012`](https://github.com/wso2/identity-apps/commit/aa6e1e6012fea5ec6c04d54f4c6e6b17e268239e)]:
+  - @wso2is/theme@2.0.17
+
+## 2.1.6
+
+### Patch Changes
+
+- [#4268](https://github.com/wso2/identity-apps/pull/4268) [`9dd93eb4fa`](https://github.com/wso2/identity-apps/commit/9dd93eb4fa45da45e5dcd87babb58e4b10afdbd8) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - Add deployment config to conditionally enable remote logging and advisory banner features
+
+## 2.1.5
+
+### Patch Changes
+
+- [#4225](https://github.com/wso2/identity-apps/pull/4225) [`583e0aa16a`](https://github.com/wso2/identity-apps/commit/583e0aa16a3ff394f62baa837e8be395557b7f70) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - Address PR review comments and UI issues in remote logging
+
+* [#4253](https://github.com/wso2/identity-apps/pull/4253) [`5ba184cff3`](https://github.com/wso2/identity-apps/commit/5ba184cff36f2342ce8444ca0b12dc1d1b16aeb3) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - Fix incorrect i18n keys in assigned groups UI in roles section
+
+- [#4239](https://github.com/wso2/identity-apps/pull/4239) [`c87b504a32`](https://github.com/wso2/identity-apps/commit/c87b504a326e6d896f6bb63d9d6cb12590376ccd) Thanks [@Achintha444](https://github.com/Achintha444)! - change assign roles interfaces in the `user` and `group` sections.
+
+* [#4263](https://github.com/wso2/identity-apps/pull/4263) [`ed53722390`](https://github.com/wso2/identity-apps/commit/ed5372239055ab28d75f9bb3630870cf4c0b4905) Thanks [@DilshanSenarath](https://github.com/DilshanSenarath)! - Remove the http://schema.openid.net/2007/05/claims dialect from the new console.
+
+- [#4233](https://github.com/wso2/identity-apps/pull/4233) [`c0ea5f71e6`](https://github.com/wso2/identity-apps/commit/c0ea5f71e62471f207bca1e927eefa56cd3c224a) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - UI improvements to the user store connections section
+
+* [#4266](https://github.com/wso2/identity-apps/pull/4266) [`d94fa3f961`](https://github.com/wso2/identity-apps/commit/d94fa3f9610eb933210e930a54499bdbf097c32b) Thanks [@Achintha444](https://github.com/Achintha444)! - Add create role UI
+
+- [#4252](https://github.com/wso2/identity-apps/pull/4252) [`199144fe70`](https://github.com/wso2/identity-apps/commit/199144fe70a3a6a7167c5e97ecdd9cddec1f3d45) Thanks [@PasinduYeshan](https://github.com/PasinduYeshan)! - Add disabled user store option in bulk user import
+
+* [#4262](https://github.com/wso2/identity-apps/pull/4262) [`f6d8450d9d`](https://github.com/wso2/identity-apps/commit/f6d8450d9d7fc43138f77b9f3744d3b25aa9cdea) Thanks [@asekawa](https://github.com/asekawa)! - Fix placeholder values in Advance section of branding
+
+- [#4267](https://github.com/wso2/identity-apps/pull/4267) [`aaaa2c9118`](https://github.com/wso2/identity-apps/commit/aaaa2c91180bef44f64a77d5ef20af740dab0b53) Thanks [@savindi7](https://github.com/savindi7)! - Hide My Account from Navbar
+
+- Updated dependencies [[`583e0aa16a`](https://github.com/wso2/identity-apps/commit/583e0aa16a3ff394f62baa837e8be395557b7f70), [`ed53722390`](https://github.com/wso2/identity-apps/commit/ed5372239055ab28d75f9bb3630870cf4c0b4905), [`c0ea5f71e6`](https://github.com/wso2/identity-apps/commit/c0ea5f71e62471f207bca1e927eefa56cd3c224a), [`d94fa3f961`](https://github.com/wso2/identity-apps/commit/d94fa3f9610eb933210e930a54499bdbf097c32b), [`199144fe70`](https://github.com/wso2/identity-apps/commit/199144fe70a3a6a7167c5e97ecdd9cddec1f3d45)]:
+  - @wso2is/i18n@2.0.10
+  - @wso2is/theme@2.0.16
+
 ## 2.1.4
 
 ### Patch Changes